Ensure translate-dynamic runtime helper is namespaced.
Without this change, including
ember-holy-futuristic-template-namespacing-batman at multiple versions
would have an unknown version of this runtime helper. The AST transform
would run correctly against each "layer", but the runtime helpers are
merged with the "last one wins" strategy.

Versioning this runtime helper prevents the issue, and guarantees that
we can have v0.1 and the upcoming v0.2 running together in the same
build.