Fix dependency check of ember-resolver in project
Currently the addon is looking for ember-resolver within itself. Since
ember-resolver is most of the times (if not always) a devDependency, it
causes issues when using `npm link` or `yarn link` with addons that use
this addon.

When using `this.project`, we are confirming that the version of
`ember-checker` is correct at the project level, because it is expected
to be a project dependency.